agent/exec/dockerapi: move to swarmd/dockerexec
The Docker API executor is one of the only importers of
github.com/docker/docker and is only used by swarmd. Move it under the
swarmd package tree in preparation for turning swarmd into its own Go
module.
