+++ swarmd/dockerexec/controller_integration_test.go
@@ -1,4 +1,4 @@
-package dockerapi
+package dockerexec
 
 import (
 	"context"
