+++ swarmd/dockerexec/docker_client_stub.go
@@ -1,4 +1,4 @@
-package dockerapi
+package dockerexec
 
 import (
 	"context"
